+++ oui/oui_radio.h
@@ -11,7 +11,7 @@ class OUI_API UIRadio : public UILink
 public:
 	UIRadio();
 	UIRadioType type;
-	void on_click(int x, int y, size_t param) override;
+	void on_click(int x, int y, uint32_t param) override;
 	bool select(bool bSelect) override;
 	void on_update() override;
 };
